+++ mobile/android/gradle.properties
@@ -1,6 +1,17 @@
 org.gradle.jvmargs=-Xmx8G -XX:MaxMetaspaceSize=4G -XX:ReservedCodeCacheSize=512m -XX:+HeapDumpOnOutOfMemoryError
 android.useAndroidX=true
+
 # This builtInKotlin flag was added automatically by Flutter migrator
+
 android.builtInKotlin=false
+
 # This newDsl flag was added automatically by Flutter migrator
+
 android.newDsl=false
+
+# Desativa a compilação incremental do Kotlin para evitar erros de cache
+
+# entre o projeto no drive D: e dependências armazenadas no drive C:
+
+kotlin.incremental=false
+kotlin.incremental.android=false
