+++ mobile/lib/controller/auth/controller_reset_password.dart
@@ -3,8 +3,8 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
-import '../../service/API.dart';
-import '../../util/Translation.dart';
+import '../../service/api.dart';
+import '../../util/translation.dart';
 
 class ControllerResetPassword {
   ControllerResetPassword(this._notify, this.otp);
