+++ mobile/lib/controller/resume/controller_new_resume.dart
@@ -4,8 +4,8 @@ import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
-import '../../service/API.dart';
-import '../../util/Translation.dart';
+import '../../service/api.dart';
+import '../../util/translation.dart';
 
 class Skill {
   final TextEditingController name = TextEditingController();
