+++ mobile/lib/view/auth/view_forgot_password.dart
@@ -1,12 +1,12 @@
-import 'package:bomcurriculo/include/BodyAuth.dart';
-import 'package:bomcurriculo/theme/AppColors.dart';
-import 'package:bomcurriculo/util/Translation.dart';
-import 'package:bomcurriculo/widget/WidgetError.dart';
+import 'package:bomcurriculo/include/body_auth.dart';
+import 'package:bomcurriculo/theme/app_colors.dart';
+import 'package:bomcurriculo/util/translation.dart';
+import 'package:bomcurriculo/widget/widget_error.dart';
 import 'package:flutter/material.dart';
 
-import '../../controller/auth/ControllerForgotPassword.dart';
-import '../../widget/WidgetButton.dart';
-import '../../widget/WidgetInputText.dart';
+import '../../controller/auth/controller_forgot_password.dart';
+import '../../widget/widget_button.dart';
+import '../../widget/widget_input_text.dart';
 
 class ViewForgotPassword extends StatefulWidget {
   const ViewForgotPassword({super.key});
