+++ mobile/lib/view/auth/view_verify_otp.dart
@@ -1,12 +1,12 @@
-import 'package:bomcurriculo/include/BodyAuth.dart';
-import 'package:bomcurriculo/theme/AppColors.dart';
-import 'package:bomcurriculo/util/Translation.dart';
-import 'package:bomcurriculo/widget/WidgetError.dart';
+import 'package:bomcurriculo/include/body_auth.dart';
+import 'package:bomcurriculo/theme/app_colors.dart';
+import 'package:bomcurriculo/util/translation.dart';
+import 'package:bomcurriculo/widget/widget_error.dart';
 import 'package:flutter/material.dart';
 
-import '../../controller/auth/ControllerVerifyOTP.dart';
-import '../../widget/WidgetButton.dart';
-import '../../widget/WidgetInputText.dart';
+import '../../controller/auth/controller_verify_otp.dart';
+import '../../widget/widget_button.dart';
+import '../../widget/widget_input_text.dart';
 
 class ViewVerifyOTP extends StatefulWidget {
   const ViewVerifyOTP({super.key});
