+++ mobile/lib/widget/widget_error.dart
