+++ mobile/lib/widget/widget_logo.dart
